Bluetooth: classic: OBEX: Fix action id type issue
The OBEX packet cannot be parsed properly if the action ID header is
included in the packet.

It is caused by the value type issue of the action ID. The action ID
type should be `uint8_t` instead of `uint32_t`.

Fix the issue by change action ID value type from `uint32_t` to
`uint8_t`.

+++ subsys/bluetooth/host/classic/shell/goep.c
@@ -932,7 +932,7 @@ static int cmd_add_header_action_id(const struct shell *sh, size_t argc, char *a
 
 	action_id = strtoul(argv[1], NULL, 16);
 
-	err = bt_obex_add_header_action_id(goep_app.tx_buf, action_id);
+	err = bt_obex_add_header_action_id(goep_app.tx_buf, (uint8_t)action_id);
 	if (err) {
 		shell_error(sh, "Fail to add header action_id");
 	}
